modem: fix registered status when connected
ModemManager status is showing as Registered while it is connected.

ModemManager state is changing from CONNECTED to ENABLED state
when the modem is idle.

Introduced a check in ModemManager to see if previous state
is CONNECTED and any bearer is present, if so state is updated
to previous state (CONNECTED) or else continue with the state
change.

Signed-off-by: Anubhav Gupta <anubhav.gupta@digi.com>