QMAP: Remove newer WDA data formats
Some modems don't respond to DHCP requests when using the newer WDA
formats, for now just use the QMAP/RAW IP/QMI WWAN format, the same
as the DAL patch for multiplexing which was removed to use the built in
multiplexing support.

Ideally we can support QMAPv5, but for now this is causing issues with
Telit modems and some of the Sierra modems.