Quectel: Disable QUSIM unsolicted event for Quectel modems
With MM 1.20, an unsolicited QUSIM issued by the modem was causing MM
to enter the SIM hot swap code, which results in the modem getting unmapped
and remapped. That was adding over a minute before the modem comes up when
SIM is PIN locked. This change hides the QUSIM unsolicited event on Quectel
modems.

DAL-8189