operator: Ignore "DATA ONLY" operator name
When the modem is in a roaming state, it has been observed to return
"DATA ONLY" as the service name during PLMN name retrieval. This update
modifies the behavior to disregard the "DATA ONLY" and, instead,
utilize the operator name if available.

This is sample of the logs when "DATA ONLY" is returned.

[modem0] current operator long name: vodafone UK
[modem0] current operator short name: voda UK
[modem0] current operator service name: DATA ONLY

DAL-9315