Cinterion plugin: Removed ID_MM_PORT_TYPE_AT_PRIMARY tag for MV31-W in MBIM mode PID.
When ModemManager detects and creates te MV31-W (with the model name not displayed),
it occasionally fails to find the AP port. This issue occurs because, when the
Cinterion plugin is detected, its custom_init function is called and the function
checks if the port is already tagged as an AT port (This was added in mm-1.22).
If it is, probing stops, and the plugin begins using the AT port but the
AT port might not yet be ready.

Removing the ID_MM_PORT_TYPE_AT_PRIMARY tag allows the Cinterion plugin's
custom_init function and ModemManager to continue probing the AT port.