mm-bearer-mbim: add session-id bearer property
This commit adds support for a new 'session-id' property in the bearer
configuration.  If provided, the specified session-id is used in the
MBIM connect message, allowing explicit control (including zero) over
the session id used for MBIM connections.  This is useful for modems or
networks that require a specific session id.

DAL-10697